+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Next]
 
 ### Added
--
+- `GetCurrencyInfo()` extension methods on `CultureInfo` and `RegionInfo` for easy access to `CurrencyInfo`.
 
 ### Changed
 -
@@ -24,8 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved performance for Money creation, arithmetic, and formatting (up to 30%) by encoding `MinorUnit = 2` hint
   directly into the `Currency` struct's 15th bit for faster rounding and lookups.
 - IRoundingStrategy.Round() uses Currency as a parameter instead of CurrencyInfo.
-- Fixed IntroductionOn date for SLE currency https://github.com/RemyDuijkeren/NodaMoney/pull/117 by @General-66.
-- Compact format uses given digits as significant digits where the default is 2,
+- Fixed `IntroductionOn` date for SLE currency https://github.com/RemyDuijkeren/NodaMoney/pull/117 by @General-66.
+- Compact format use given digits as significant digits where the default is 2,
   values below 1000 are displayed as normal currency values where > 100 without decimals https://github.com/RemyDuijkeren/NodaMoney/issues/111,
 
 ### Removed